requirements.txt
Django >=2.2.6,<2.3.0
djangorestframework>=3.9.2,<3.10.0
psycopg2>=2.8.4,<2.9.0
Pillow>=6.2.1,<7
flake8>=3.7.8,<3.8.0
django-cors-headers>=3.0.0,<3.2.2
django-extensions
ez_setup
cryptography
django-rest-knox
django-filter